create a new interface for user and userfactory, and implemented the CommonUser and CommonUserFactory
