+++ src/entity/UserFactory.java
@@ -1,0 +1,5 @@
+package entity;
+
+public interface UserFactory {
+    CommonUser create(String username, String password);
+}
