--- src/entity/UserLogin.java
@@ -1,19 +1,0 @@
-package entity;
-
-public class UserLogin {
-    private final String username;
-    private final String password;
-
-    public UserLogin(String username, String password) {
-        this.username = username;
-        this.password = password;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
